chore(www): reorganize
