+++ www/components/Layout/ProfileButton/ProfileButton.js
@@ -9,7 +9,7 @@ import {
 import { AccountCircle, Logout } from '@mui/icons-material';
 import { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { userLogout } from '../../store/features/user';
+import { userLogout } from '../../../store/features/user';
 
 export default function ProfileButton() {
   const [ anchor, setAnchor ] = useState( null );
