--- www/components/StatusButton/StatusButton.module.scss
@@ -1,25 +1,0 @@
-@use '../../styles/colors.scss' as colors;
-
-.action_button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  p {
-    margin: 0;
-    padding: 0;
-    margin-right: 2em;
-  }
-
-  .arrow {
-    color: colors.$midText;
-  }
-
-  .first_tag {
-    margin-right: .25em;
-  }
-
-  .second_tag {
-    margin-left: .25em;
-  }
-}
