+++ www/components/shared/LoadingIcon/LoadingIcon.js
