+++ www/components/shared/Tag/Tag.module.scss
@@ -1,4 +1,4 @@
-@use '../../styles/colors' as colors;
+@use '../../../styles/colors' as colors;
 
 .tag {
   color: colors.$lightText;
