+++ www/pages/register.js
@@ -48,7 +48,7 @@ const Register = props => {
 
         const user = props.store.getState().user;
 
-        router.push( `/user/${ user._id }` );
+        router.push( `/user/${ user._id }/home` );
 
       } catch ( err ) {
         await props.store.dispatch(
