+++ www/styles/globals.css
@@ -59,7 +59,6 @@ p {
 /* THEMES */
 
 .default {
-
   --primary: #3A3C45;
   --primary-hover: #1d1e23;
   --background: #f5f5f5;
